requirements.txt
streamlit
groq
pydantic
python-pptx
pandas
openpyxl
requests